requirements.txt
numpy==1.21.4
opencv_python==4.5.5.64
Pillow==9.2.0
streamlit==1.11.0
tensorflow==2.9.1
tensorflow_gpu==2.8.0
